==> script/01_coordenadas_municipios.ipynb <==
### Instalar pacotes se necessário

#pip install pandas geopandas shapely

### Carregar pacotes

import pandas as pd
import geopandas as gpd
from shapely import wkt

# Obter coordenadas centrais dos municípios brasileiros

### Carregar arquivo da malha territorial dos municípios

territorios = gpd.read_file(
  "../datasets/BR_Municipios_2021/BR_Municipios_2021.shp")
territorios.crs

### Calcular os centroides dos municípios

territorios['pontos'] = territorios.centroid
territorios.head()
#ignorar o aviso: a reprojeção é desnecessária, pois os centroides aproximados são adequados para
#a análise proposta, sendo necessário apenas um ponto representativo dos municípios brasileiros.

### Plotar a malha territorial e os centroides dos municípios para verificação da adequação

territorios.plot()
territorios['pontos'].plot()

### Obter as coordenadas dos centroides como strings

pontos = territorios['pontos'].apply(lambda x: wkt.dumps(x))
print(type(pontos[0]))
pontos.head()

### Remover informações desnecessárias das strings

pontos = [i.replace('POINT (','') for i in pontos]
pontos = [i.replace(')','') for i in pontos]
pontos[0:5]

### Separar os dados de long e lat e inserir em um data frame

pontos = pd.DataFrame(pontos)[0].str.split(' ',n=1,expand=True)
pontos.head()

### Juntar as tabelas com as coordenadas e nomes dos municípios e estado

pontos = pontos.join(territorios[['NM_MUN','SIGLA']])
pontos.columns = ['long','lat','Municipio','Estado']
pontos

### Salvar a tabela em arquivo

pontos.to_csv('../dados_coletados/municipios_coordenadas.csv', index = False)
