==> script/02_dados_agricola.ipynb <==
import pandas as pd
import numpy as np

# Dados de produção de soja por município

### Carregar os dados

dados_original = pd.read_csv("../datasets/soja_1990_2005.csv", na_values='..')
dados_original.head()

### Renomear colunas

dados_original.columns = ['Municipio', 'Ano', 'Variavel', 'Produto', 'Valor']
dados_original.head()

### Verificar atributos e transpor

print(dados_original['Produto'].unique())
#manter apenas as linhas referentes à produção de soja
dados_soja = dados_original[dados_original['Produto'] == 'Soja (em grão)']
display(dados_soja)

# Manter apenas uma combinação de Municipio e Ano
tabela_soja = dados_soja[['Municipio','Ano']].drop_duplicates()
tabela_soja

#Selecionar atributos que serão transpostos para a tabela
atributos = dados_soja['Variavel'].unique()
atributos

for i in atributos:
    variavel = dados_soja[dados_soja['Variavel'] == i]
    variavel = variavel.rename(columns={'Valor': i})
    variavel = variavel.drop(['Variavel','Produto'], axis=1)
    tabela_soja = tabela_soja.merge(variavel, on=['Municipio','Ano'])
tabela_soja.tail()

### Separar a sigla do estado em uma coluna distinta

split = tabela_soja["Municipio"].str.split("(", n = 1, expand = True)
tabela_soja["Municipio"] = split[0]
tabela_soja["Estado"] = split[1]
tabela_soja['Estado'] = tabela_soja['Estado'].str. replace(')','', regex = False)
tabela_soja['Municipio'] = tabela_soja['Municipio'].str.strip()
tabela_soja.head()

tabela_soja.columns = [
    'Municipio','Ano','Area_total','Area_%','Area_colhida',
    'A_colhida_%','Toneladas','Rendimento','Estado'
]
tabela_soja = tabela_soja.reindex(columns = [
    'Municipio','Ano','Estado','Area_total','Area_%',
    'Area_colhida','A_colhida_%','Toneladas','Rendimento'
])
tabela_soja.tail()

### Savar tabela em arquivo

tabela_soja.to_csv('../dados_coletados/tabela_soja.csv', index = False)
