==> script/07_tabela_futuro.ipynb <==
### Configuração de pacotes

import pandas as pd
from functools import reduce
import glob
import os

### Juntar os dados das variáveis históricas em uma única tabela

#### Definir método para os cenários e tempos futuros

#Cenário da mudança climática: otimista = 'CE0007', intenso = 'CE0009'
#Período temporal: 2011-2040 = 'PE0001', 2041-2070 = 'PE0002', 2071-2100 = 'PE0003'
def tabela_clima(cenario, tempo):
    arquivos = glob.glob(os.path.join('../dados_coletados/futuro/' + 
                                      cenario + '/' + tempo + '/','*.csv'))
    print(len(arquivos))
    lista = []
    for arquivo in arquivos:
        tabela = pd.read_csv(arquivo)
        lista.append(tabela)
    print(len(lista))
    futuro = reduce(
        lambda tabela1, tabela2:
        pd.merge(tabela1, tabela2,
             on = ['Estado','Municipio','Ano']),
        lista)
    display(futuro.head())
    futuro.to_csv('../dados_coletados/' + cenario + '_' + tempo + '.csv', index = False)

##### Cenário otimista (RCP4.5) de 2041 a 2070

tabela_clima('CE0007', 'PE0002')

##### Cenário otimista (RCP4.5) de 2071 a 2100

tabela_clima('CE0007', 'PE0003')

##### Cenário intenso (RCP8.5) de 2041 a 2070

tabela_clima('CE0009', 'PE0002')

##### Cenário intenso (RCP8.5) de 2071 a 2100

tabela_clima('CE0009', 'PE0003')
