==> script/08_tabela_modelagem.ipynb <==
import pandas as pd

### Juntar dados agrícolas e dados climáticos para modelagem

dados = pd.read_csv("../dados_coletados/dados.csv")
historico = pd.read_csv('../dados_coletados/historico.csv')

tabela_modelagem = pd.merge(dados, historico, on = ['Estado','Municipio','Ano'])

tabela_modelagem[tabela_modelagem['Rendimento'].isnull()]

tabela_modelagem = tabela_modelagem.drop(
    tabela_modelagem[tabela_modelagem['Rendimento'].isnull()].index
)

tabela_modelagem

tabela_modelagem = tabela_modelagem.drop(
    tabela_modelagem[tabela_modelagem['Rendimento'] == '-'].index
)

tabela_modelagem['Rendimento'] = tabela_modelagem['Rendimento'].astype(float)
print(tabela_modelagem['Rendimento'].dtypes)

tabela_modelagem

tabela_modelagem.to_csv('../dados_coletados/tabela_modelagem.csv', index = False)
